#74 cleanup + NET48 variant of ToolkitApiController
