+++ KX13.Extensions/KX13.Extensions.csproj
@@ -1,0 +1,14 @@
+<Project Sdk="Microsoft.NET.Sdk">    
+    <PropertyGroup>        
+        <Nullable>disable</Nullable>
+        <TargetFrameworks>net5.0;net6.0;netcoreapp3.1</TargetFrameworks>
+        <LangVersion>8</LangVersion>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Kentico.Xperience.AspNetCore.WebApp" Version="13.0.73" />
+      <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.2.0" />
+      <PackageReference Include="Microsoft.AspNetCore.Mvc.Abstractions" Version="2.2.0" />
+    </ItemGroup>
+
+</Project>
